// packages/codex-orchestration-cli/hooks/generated/pre-tool-use-output.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type PreToolUseDecisionWire = "approve" | "block";
export type PreToolUsePermissionDecisionWire = "allow" | "deny" | "ask";

export interface PreToolUseCommandOutput {
  continue?: boolean;
  decision?: PreToolUseDecisionWire;
  hookSpecificOutput?: PreToolUseHookSpecificOutputWire;
  reason?: string;
  stopReason?: string;
  suppressOutput?: boolean;
  systemMessage?: string;
}
export interface PreToolUseHookSpecificOutputWire {
  additionalContext?: string;
  hookEventName: "PreToolUse";
  permissionDecision?: PreToolUsePermissionDecisionWire;
  permissionDecisionReason?: string;
  updatedInput?: {
    [k: string]: unknown;
  };
}