// packages/codex-orchestration-cli/hooks/generated/session-start-output.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SessionStartCommandOutput {
  continue?: boolean;
  hookSpecificOutput?: SessionStartHookSpecificOutputWire;
  stopReason?: string;
  suppressOutput?: boolean;
  systemMessage?: string;
}
export interface SessionStartHookSpecificOutputWire {
  additionalContext?: string;
  hookEventName: "SessionStart";
}